Download aosp images (#92)
* Update to python3

python3 is available on all supported debian and ubuntu distros.
python2.7 is no longer supported at google so we are switching to
python3.

* cvd_container_exists lists inactive containers


* download-aosp.sh: add options to stream/uncompress and skip userdata

-A and/or -C will now stream and uncompress images on the fly rather
    than first downloading them, to save on disk;
-U in combination with -A will now skip downloading userdata.img; this
    will require passing --data_policy=create_if_missing and
    --blank_data_image_mb=4096MB.
-D switches to the previous beavior (by default off) in case there are
   network issues and it's better to download the whole archives first.
-K preserves the downloaded archives (so it requires -D) rather than
   deleting them.
-X does not extract kept archives (so it requires -K and -D).

If you just want to download the latest AOSP Android and Cuttlefish
images:
    ./download-aosp.sh -AC -DKX -a $(uname -m)

If you want to save space and stream-uncompress Android and Cuttlefish
while skipping userdata:
    ./download-aosp.sh -AC -U -a $(uname -m)

The streaming optoin uses bsdtar to stream images and exclude
userdata.img.

+++ download-aosp.sh
@@ -7,6 +7,10 @@ function print_help {
 	echo "     Options:"
 	echo "       -a | --arch                      : architecture to download for: x86, x86_64, arm64"
 	echo "       -A | --android                   : download the Android OS images (by default disabled)"
+	echo "       -D | --download (default off)    : download images rather than streaming/uncompressing on the fly"
+	echo "       -K | --keep (default off)        : keep archives after extracting them (requires -D)"
+	echo "       -X | --no-extract (default off)   : do not extract extract downloaded archives (requires -D -K)"
+	echo "       -U | --no-userdata (default off) : skip downloading userdata.img (requires -A and not -D)"
 	echo "       -C | --cuttlefish                : download the Cuttlefish host images (by default disabled)"
 	echo "       -h | --help                      : print this help message"
 }
@@ -15,7 +19,7 @@ function parse_opts() {
 	local help=0
 	local params
 
-	if params=$(getopt -o 'a:ACh' -l 'arch:,android,cuttlefish,help' --name "$0" -- "$@"); then
+	if params=$(getopt -o 'a:ADKXUCh' -l 'arch:,android,download,keep,no-userdata,cuttlefish,help' --name "$0" -- "$@"); then
 		eval set -- "${params}"
 		while true; do
 			case "$1" in
@@ -47,6 +51,22 @@ function parse_opts() {
 					DOWNLOAD_CF=1
 					shift
 					;;
+				-D|--download)
+					DOWNLOAD=1
+					shift
+					;;
+				-K|--keep)
+					KEEP=1
+					shift
+					;;
+				-X|--no-extract)
+					EXTRACT=0
+					shift
+					;;
+				-U|--no-userdata)
+					SKIP_USERDATA='--exclude userdata.img'
+					shift
+					;;
 				-h|--help)
 					help=1
 					shift
@@ -69,14 +89,48 @@ function parse_opts() {
 		help=1
 	fi
 
+	if [[ -n "${SKIP_USERDATA}" && "${DOWNLOAD_ANDROID}" -eq 0 ]]; then
+		echo "Option -U/--no-userdata requires -A/--android" 1>&2
+		help=1
+	fi
+
+	if [[ -n "${SKIP_USERDATA}" && "${DOWNLOAD}" -eq 1 ]]; then
+		echo "Option -U/--no-userdata does not work with -D/--download" 1>&2
+		help=1
+	fi
+
+	if [[ "${KEEP}" -eq 1 && "${DOWNLOAD}" -eq 0 ]]; then
+		echo "Option -K/--keep requires -D/--download" 1>&2
+		help=1
+	fi
+
+	if [[ "${EXTRACT}" -eq 0 && "${KEEP}" -eq 0 ]]; then
+		echo "Option -X/--no-extract requires -K/--keep" 1>&2
+		help=1
+	fi
+
 	if [ "${help}" -eq 1 ]; then
 		print_help
 		exit 1
+	fi
+
+	if [[ -n "${SKIP_USERDATA}" ]]; then
+cat<<EOF
+
+    NOTE: A blank userdata.img will not be downloaded.  Make sure to start
+          Cuttlefish with the options --data_policy create_if_missing
+          --blank_data_image_mb=<size in MB>.
+
+EOF
 	fi
 }
 
 DOWNLOAD_ANDROID=0
 DOWNLOAD_CF=0
+DOWNLOAD=0
+KEEP=0
+EXTRACT=1
+SKIP_USERDATA=""
 
 parse_opts $*
 
@@ -89,9 +143,17 @@ if [ "${DOWNLOAD_ANDROID}" -eq 1 ]; then
 	echo "#"
 	set -x
 	IMG=aosp_cf_${ARCH}_phone-img-$(echo $RURL | awk -F\/ '{print $6}').zip
-	wget -nv ${RURL%/view/BUILD_INFO}/raw/${IMG}
-	unzip "${IMG}"
-	rm -v "${IMG}"
+	if [[ "${DOWNLOAD}" -eq 1 ]]; then
+		wget -nv ${RURL%/view/BUILD_INFO}/raw/${IMG}
+		if [[ "${EXTRACT}" -eq 1 ]]; then
+			unzip "${IMG}"
+		fi
+		if [[ "${KEEP}" -eq 0 ]]; then
+			rm -v "${IMG}"
+		fi
+	else
+		wget -O- -nv --show-progress --progress=bar:giga --prefer-family=IPv4 -c ${RURL%/view/BUILD_INFO}/raw/${IMG} | bsdtar -x -v -f- ${SKIP_USERDATA}
+	fi
 	set +x
 fi
 
@@ -100,8 +162,16 @@ if [ "${DOWNLOAD_CF}" -eq 1 ]; then
 	echo "# Downloading the latest Cuttlefish images from AOSP"
 	echo "#"
 	set -x
-	wget -nv ${RURL%/view/BUILD_INFO}/raw/cvd-host_package.tar.gz
-	tar xzvf cvd-host_package.tar.gz
-	rm -v cvd-host_package.tar.gz
+	if [[ "${DOWNLOAD}" -eq 1 ]]; then
+		wget -nv ${RURL%/view/BUILD_INFO}/raw/cvd-host_package.tar.gz
+		if [[ "${EXTRACT}" -eq 1 ]]; then
+			tar xzvf cvd-host_package.tar.gz
+		fi
+		if [[ "${KEEP}" -eq 0 ]]; then
+			rm -v cvd-host_package.tar.gz
+		fi
+	else
+		wget -O- -nv --show-progress --progress=bar:giga --prefer-family=IPv4 -c ${RURL%/view/BUILD_INFO}/raw/cvd-host_package.tar.gz | tar -xzvf-
+	fi
 	set +x
 fi
